feat: use auto-generated client
